Lavaland Goidamerge (#3644)
* Goidamerge method & refactoring of LavalandPlanetSystem

* chore: Automatically update REUSE headers

* chore: Automatically update REUSE headers

* fix things

* Fix maps

---------
