+++ Content.IntegrationTests/Tests/_Lavaland/LavalandGenerationTest.cs
@@ -8,6 +8,7 @@
 // SPDX-FileCopyrightText: 2025 Piras314 <p1r4s@proton.me>
 // SPDX-FileCopyrightText: 2025 Rouden <149893554+Roudenn@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Roudenn <romabond091@gmail.com>
+// SPDX-FileCopyrightText: 2025 SX-7 <sn1.test.preria.2002@gmail.com>
 // SPDX-FileCopyrightText: 2025 TheBorzoiMustConsume <197824988+TheBorzoiMustConsume@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Unlumination <144041835+Unlumy@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
@@ -32,7 +33,7 @@ using Robust.Shared.GameObjects;
 namespace Content.IntegrationTests.Tests._Lavaland;
 
 [TestFixture]
-[TestOf(typeof(LavalandPlanetSystem))]
+[TestOf(typeof(LavalandSystem))]
 public sealed class LavalandGenerationTest
 {
     [Test]
@@ -46,7 +47,7 @@ public sealed class LavalandGenerationTest
         var mapMan = server.MapMan;
 
         var ticker = server.System<GameTicker>();
-        var lavaSystem = entMan.System<LavalandPlanetSystem>();
+        var lavaSystem = entMan.System<LavalandSystem>();
         var mapSystem = entMan.System<SharedMapSystem>();
 
         // Setup
