+++ Content.Server/_Lavaland/Procedural/Components/LavalandPreloaderComponent.cs
@@ -1,5 +1,8 @@
 // SPDX-FileCopyrightText: 2025 Aiden <28298836+Aidenkrz@users.noreply.github.com>
+// SPDX-FileCopyrightText: 2025 GoobBot <uristmchands@proton.me>
 // SPDX-FileCopyrightText: 2025 Rouden <149893554+Roudenn@users.noreply.github.com>
+// SPDX-FileCopyrightText: 2025 Roudenn <romabond091@gmail.com>
+// SPDX-FileCopyrightText: 2025 SX-7 <sn1.test.preria.2002@gmail.com>
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
@@ -7,7 +10,7 @@ using Content.Server._Lavaland.Procedural.Systems;
 
 namespace Content.Server._Lavaland.Procedural.Components;
 
-[RegisterComponent, Access(typeof(LavalandPlanetSystem))]
+[RegisterComponent, Access(typeof(LavalandSystem))]
 public sealed partial class LavalandPreloaderComponent : Component
 {
