+++ Content.Server/_Lavaland/Weather/Gamerule/LavalandStormSchedulerRule.cs
@@ -8,6 +8,7 @@
 // SPDX-FileCopyrightText: 2025 Piras314 <p1r4s@proton.me>
 // SPDX-FileCopyrightText: 2025 Rouden <149893554+Roudenn@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Roudenn <romabond091@gmail.com>
+// SPDX-FileCopyrightText: 2025 SX-7 <sn1.test.preria.2002@gmail.com>
 // SPDX-FileCopyrightText: 2025 TheBorzoiMustConsume <197824988+TheBorzoiMustConsume@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Unlumination <144041835+Unlumy@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
@@ -36,7 +37,7 @@ public sealed class LavalandStormSchedulerRule : GameRuleSystem<LavalandStormSch
     [Dependency] private readonly IPrototypeManager _proto = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
     [Dependency] private readonly IChatManager _chatManager = default!;
-    [Dependency] private readonly LavalandPlanetSystem _lavaland = default!;
+    [Dependency] private readonly LavalandSystem _lavaland = default!;
     [Dependency] private readonly LavalandWeatherSystem _lavalandWeather = default!;
 
     public override void Update(float frameTime)
