+++ Content.Shared/_Lavaland/Procedural/Prototypes/LavalandGridRuinPrototype.cs
@@ -8,6 +8,7 @@
 // SPDX-FileCopyrightText: 2025 Piras314 <p1r4s@proton.me>
 // SPDX-FileCopyrightText: 2025 Rouden <149893554+Roudenn@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Roudenn <romabond091@gmail.com>
+// SPDX-FileCopyrightText: 2025 SX-7 <sn1.test.preria.2002@gmail.com>
 // SPDX-FileCopyrightText: 2025 TheBorzoiMustConsume <197824988+TheBorzoiMustConsume@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 Unlumination <144041835+Unlumy@users.noreply.github.com>
 // SPDX-FileCopyrightText: 2025 coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
@@ -36,10 +37,13 @@ public sealed partial class LavalandGridRuinPrototype : IPrototype
     [DataField] public LocId Name = "lavaland-ruin-unknown";
 
     [DataField(required: true)]
-    public ResPath Path { get; } = default!;
+    public ResPath Path { get; }
 
     [DataField]
     public int SpawnAttemps = 8;
+
+    [DataField]
+    public bool PatchToPlanet = true;
 
     [DataField(required: true)]
     public int Priority = int.MinValue;
