+++ Resources/Prototypes/_Lavaland/Procedural/biome_templates.yml
@@ -28,7 +28,7 @@
         frequency: 1
         seed: 15
       allowedTiles:
-        - FloorBasalt
+        - FloorBasaltLavaland
       entities:
         - BasaltOne
         - BasaltTwo
@@ -37,7 +37,7 @@
         - BasaltFive
     - !type:BiomeDecalLayer
       allowedTiles:
-        - FloorBasalt
+        - FloorBasaltLavaland
       threshold: 0.95
       divisions: 1
       noise:
@@ -60,7 +60,7 @@
         noiseType: OpenSimplex2
         frequency: 2
       allowedTiles:
-        - FloorBasalt
+        - FloorBasaltLavaland
       entities:
         - FloraRockSolid
     - !type:BiomeEntityLayer
@@ -73,13 +73,13 @@
         lacunarity: 2
         gain: 0.4
       allowedTiles:
-        - FloorBasalt
+        - FloorBasaltLavaland
       entities:
         - FloorLavaEntity
     # Rock formations
     - !type:BiomeEntityLayer
       allowedTiles:
-        - FloorBasalt
+        - FloorBasaltLavaland
       threshold: -0.25
       noise:
         seed: 5
@@ -97,4 +97,4 @@
     # Fill basalt
     - !type:BiomeTileLayer
       threshold: -1
-      tile: FloorBasalt
+      tile: FloorBasaltLavaland
