+++ Resources/Prototypes/_Lavaland/Procedural/lavaland_planets.yml
@@ -42,7 +42,6 @@
   markers:
   - RandomTendrilLavaland
   - RandomMobLavaland
-  - ChasmLavaland
   - OreIron
   - OreCoal
   - OreQuartz
