+++ Resources/Prototypes/_Lavaland/Tiles/hiero_arena.yml
@@ -21,7 +21,7 @@
 
 - type: tile
   id: LavalandFloorHierophantArena
-  name: tiles-basalt-floor
+  name: DO NOT MAP
   sprite: /Textures/Tiles/Planet/basalt.png
   isSubfloor: false
   footstepSounds:
