+++ Resources/Prototypes/_Lavaland/Tiles/planet.yml
@@ -1,0 +1,11 @@
+# Lava
+- type: tile
+  id: FloorBasaltLavaland
+  name: basalt floor [Planetmap]
+  sprite: /Textures/Tiles/Planet/basalt.png
+  isSubfloor: true
+  footstepSounds:
+    collection: FootstepAsteroid
+  heatCapacity: 10000
+  weather: true
+  indestructible: true
